+++ lib/cubes.dart
@@ -2,10 +2,12 @@ library cube;
 
 export 'package:cubes/src/cube.dart';
 export 'package:cubes/src/cube_builder.dart';
-export 'package:cubes/src/cube_widget.dart';
 export 'package:cubes/src/injector.dart';
 export 'package:cubes/src/localization/cubes_localization_delegate.dart';
 export 'package:cubes/src/localization/strings_location.dart';
 export 'package:cubes/src/observable/observable_list.dart';
 export 'package:cubes/src/observable/observable_value.dart';
 export 'package:cubes/src/util/extensions/ext.dart';
+export 'package:cubes/src/widgets/animated_list_cube.dart';
+export 'package:cubes/src/widgets/cube_widget.dart';
+export 'package:cubes/src/widgets/cube_widget_animation.dart';